Beanを変更（@AllArgsConstructorの追加等）・一部Repsitoryで使用しているクエリの変更

+++ src/main/java/jp/ac/chitose/tms/Repositoy/ITestRecordReposiory.java
@@ -3,6 +3,6 @@ package jp.ac.chitose.tms.Repositoy;
 import jp.ac.chitose.tms.Bean.TestRecordItem;
 
 public interface ITestRecordReposiory {
-	public TestRecordItem selectLatestTestRecord(long testId);
+	public TestRecordItem selectLatestTestRecord(Integer testId);
 
 }
