+++ src/main/java/jp/ac/chitose/tms/Service/ITestRecordService.java
@@ -3,5 +3,5 @@ package jp.ac.chitose.tms.Service;
 import jp.ac.chitose.tms.Bean.TestProgressItem;
 
 public interface ITestRecordService {
-	public TestProgressItem getTestProgress(long productId);
+	public TestProgressItem getTestProgress(Integer productId);
 }
